adding a load test for insering given number of wildfly servers into inventory
