+++ hawkular-inventory-load-tests/src/test/resources/gatling.conf
@@ -1,5 +1,5 @@
 #
-# Copyright 2015 Red Hat, Inc. and/or its affiliates
+# Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
 # and other contributors as indicated by the @author tags.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,7 +16,10 @@
 #
 
 gatling {
+  http {
+    enableGA = false # Google Analytics
+  }
   data {
-    writers = "file"
+    writers = "file" # The list of DataWriters to which Gatling write simulation data (currently supported : console, file, graphite, jdbc)
   }
 }
