+++ hawkular-inventory-load-tests/src/test/scala/org/hawkular/inventory/loadtest/AgentSimulation.scala
@@ -22,47 +22,7 @@ import io.gatling.core.Predef._
 import io.gatling.http.Predef._
 import io.gatling.jdbc.Predef._
 
-/**
- * All the curl commands that are commented out in the code should pass
- */
-class AgentSimulation extends Simulation {
-
-  // --------------------------- Options
-
-  val baseURI: String = System.getProperty("ltest.baseURI", "http://localhost:8080")
-  val username: String = System.getProperty("ltest.username", "jdoe")
-  val password: String = System.getProperty("ltest.password", "password")
-  val logLevel: Int = Integer.getInteger("ltest.logLevel", 0)
-
-  // Number of concurrent clients
-  val users: Int = Integer.getInteger("ltest.users", 7)
-
-  // Delay before firing up another client
-  val ramp: Long = java.lang.Long.getLong("ltest.ramp", 5L)
-
-  // The number of resource types for each client
-  val resourceTypesNumber: Int = Integer.getInteger("ltest.resourceTypes", 2)
-
-  // The number of metric types for each client
-  val metricTypesNumber: Int = Integer.getInteger("ltest.metricTypes", resourceTypesNumber * 2)
-
-  // The number of resources for each client
-  val resourcesNumber: Int = Integer.getInteger("ltest.resources", 2)
-
-  // The number of metrics for each client
-  val metricsNumber: Int = Integer.getInteger("ltest.metrics", resourcesNumber * 2)
-
-  // The number reads fo perform for each entity for each client
-  val readEntityNumber: Int = Integer.getInteger("ltest.readEntity", 2)
-
-  // Interval between requests that don't have to be synchronized (in millis)
-  val interval: Int = Integer.getInteger("ltest.interval", 1)
-
-  // ---------------------------
-
-  def log(lvl: Int, str: String): Unit = if (logLevel >= lvl) println(str) else ()
-  def log(str: String): Unit = log(1, str)
-  def logd(str: String): Unit = log(2, str)
+class AgentSimulation extends Simulation with Helpers {
 
   // http://git.io/vqEnX
   val httpConf = http
@@ -86,62 +46,6 @@ class AgentSimulation extends Simulation {
       List(info.response.bodyLength)
     })
 
-  val random = new util.Random
-
-//curl -ivX POST -H "Content-Type: application/json;charset=utf-8" -d '{"id": "footype", "properties": {"foo-name": "jmeno"}}' http://jdoe:password@127.0.0.1:8080/hawkular/inventory/resourceTypes
-  def resourceTypeJson(id: String): String = s"""
-    {
-      "id": "resType-$id",
-      "properties": {
-        "foo-name": "name-$id"
-      }
-    }
-  """
-
-//curl -ivX POST -H "Content-Type: application/json;charset=utf-8" -d '{"id": "aaa", "unit": "BYTES", "type": "GAUGE", "collectionInterval": 30, "properties": {"a": "b"}}' http://jdoe:password@127.0.0.1:8080/hawkular/inventory/metricTypes
-  def metricTypeJson(id: String): String = s"""
-    {
-      "id": "metricType-$id",
-      "unit": "BYTES",
-      "type": "GAUGE",
-      "collectionInterval": 30
-    }
-  """
-
-//curl -ivX POST -H "Content-Type: application/json;charset=utf-8" -d '["/aaa"]' http://jdoe:password@127.0.0.1:8080/hawkular/inventory/resourceTypes/footype/metricTypes
-  def associateTypesJson(id: String): String = s"""
-    ["/metricType-$id"]
-  """
-
-//curl -ivX POST -H "Content-Type: application/json" -H "Accept: application/json" -d '{"id": "foobar", "resourceTypePath": "../URL"}' http://jdoe:password@127.0.0.1:8080/hawkular/inventory/test/resources | less
-  def resourceJson(id: String): String = s"""
-    {
-      "id": "res-$id",
-      "resourceTypePath": "../resType-something",
-      "properties": {
-        "foo": "bar",
-        "bar": "foo"
-      }
-    }
-  """
-
-//curl -ivX POST -H "Accept: application/json" -H "Content-Type: application/json" -d '{"id": "fooMetric", "metricTypePath": "../status.code.type"}' 'http://jdoe:password@127.0.0.1:8080/hawkular/inventory/test/metrics'
-  def metricJson(id: String): String = s"""
-    {
-      "id": "metric-$id",
-      "metricTypePath": "../status.code.type",
-      "properties": {
-        "bla": "blabla",
-        "kus": "kuskus"
-      }
-    }
-  """
-
-//curl -ivX POST -H "Content-Type: application/json;charset=utf-8" -d '["../fooMetric"]' http://jdoe:password@127.0.0.1:8080/hawkular/inventory/test/resources/foobar/metrics
-  def associateMetricResourceJson(id: String): String = s"""
-    ["../metric-$id"]
-  """
-
   def getTenant = http("preparation - get tenant")
       .get("tenant")
       .check(status is 200)
@@ -149,138 +53,29 @@ class AgentSimulation extends Simulation {
   def createSomeResourceType = http("preparation - create res type")
       // create some resource type (creating resources with type URL leads pinger to glut the server log)
       .post("resourceTypes")
-      .body(StringBody(session => resourceTypeJson("something")))
+      .body(StringBody(session => FakeDataScenarios.resourceTypeJson("something")))
       .check(status.in(List(201, 409))
     )
 
-  def agentSimulation(types: Boolean = true) = {
-    // if types is true it creates `resourceTypesNumber` resource types and for each it creates `metricTypesNumber` metric types and associates the two together
-    // if types is false it does the same thing for resources and metrics
-
-    // hellpers
-    def generateRandomString = random.nextLong.toString
-    def jsonHelper(getXTypeJson: (String => String), getXJson: (String => String), key: String) =
-      StringBody(session => {
-        val id = session(key).as[String] 
-        if (types) getXTypeJson(id) else getXJson(id)
-      })
-    val typesStr = if (types) " type" else ""
-
-    exec(session => {
+  val setupScenario = scenario("Setup")
+    .exec(session => {
       logd(s"params:\nresourceTypes  ...  $resourceTypesNumber")
       logd(s"metricTypes    ...  $metricTypesNumber")
       logd(s"resources      ...  $resourcesNumber")
-      logd(s"metrics        ...  $metricsNumber\n")
+      logd(s"metrics        ...  $metricsNumber")
+      logd(s"users            ...  $users")
+      logd(s"wildflyNumber    ...  $wildflyNumber\n")
       session
     })
-    .repeat(if (types) resourceTypesNumber else resourcesNumber) {
-      exec(session => {
-        // storing the generated resource (type) id into session
-        session.set("key1", generateRandomString)
-      })
-      .exec(http(s"Creating resource$typesStr")
-        .post(if (types) "resourceTypes" else "test/resources")
-        .body(jsonHelper(resourceTypeJson, resourceJson, "key1"))
-        //.check(status.in(200 to 304))
-        .check(status is 201)
-      )
-      .pause(interval millis)
-      .repeat(if (types) metricTypesNumber else metricsNumber) {
-
-        exec(session => 
-          // storing the generated metric (type) id into session
-          session.set("key2", generateRandomString)
-        )
-        .exec(http(s"Creating metric$typesStr")
-          .post(if (types) "metricTypes" else "test/metrics")
-          .body(jsonHelper(metricTypeJson, metricJson, "key2"))
-          .check(status is 201)
-        )
-        .pause(interval millis)
-        .exec(http(s"Assoc. metric$typesStr to resource$typesStr")
-          .post(if (types) "resourceTypes/resType-${key1}/metricTypes" else "test/resources/res-${key1}/metrics")
-          .body(jsonHelper(associateTypesJson, associateMetricResourceJson, "key2"))
-          .check(status is 204)
-        )
-        .pause(interval millis)
-      }
-      .exec(session => {
-        logd("session after one resource-metric" + typesStr + " creation and association :" + session) 
-        session
-      })
-    }
-  }
-
-  val readSimulation = {
-    // it queries all the resource types, metric types, resources and metrics and for the whole list of the results of given entity type
-    // it subsequently perform a singe read operation `readEntityNumber` times
-    def readEntityMultipleTimes(entityName: String, pathPrefix: String = "") = {
-      val plural = entityName + "s"
-      exec(http(s"Geting all $plural")
-        .get(pathPrefix + plural)
-        .check(status is 200)
-        .check(jsonPath("$[*].id").findAll.saveAs("entityIds"))
-      ).exec(s => {
-          if (logLevel >= 2) println("\nids:\n" + s("entityIds"))
-          s
-     })
-      .pause(interval millis)
-      .foreach("${entityIds}", "entityId") {
-        repeat(readEntityNumber) {
-          exec(session => {
-            http(s"Getting one $entityName")
-              .get(pathPrefix + plural + session("entityId").as[String].replaceAll(" ", "%20"))//"/${java.net.URLEncoder.encode(entityId).toString}")
-              .check(status is 200)
-            session
-          })
-        }
-      }
-    }
-
-    def getAllResourcesOfGivenTypeMultipleTimes(resourceType: String) = {
-      repeat(readEntityNumber) {
-        exec(
-          http("Get resources of a type (a)")
-            .get(s"resourceTypes/$resourceType/resources")
-            .check(status is 200)
-        )
-      }
-      .pause(interval millis)
-      .repeat(readEntityNumber) {
-        exec(
-          http("Get resources of a type (b)")
-           .get("test/resources")
-           .queryParam("typeId", resourceType)
-           .check(status is 200)
-        )
-      }
-    }
-
-    readEntityMultipleTimes("resourceType")
-    .exec(readEntityMultipleTimes("metricType"))
-    .exec(readEntityMultipleTimes("resource", "test/"))
-    .exec(readEntityMultipleTimes("metric", "test/"))
-    .exec(getAllResourcesOfGivenTypeMultipleTimes("resType-something"))
-  }
-
-  val setupScenario = scenario("Setup")
     .exec(getTenant)
     .pause(1 seconds)
     .exec(createSomeResourceType)
 
-  val scenario1 = scenario("AgentSimulation (fill the inventory)")
-    .pause(4 seconds)
-    .exec(agentSimulation(true))
-    .exec(agentSimulation(false))
-
-  val scenario2 = scenario("Reads")
-    .exec(readSimulation)
-
-  val allScenarios = scenario1.exec(scenario2)
+  val loadTests = NTimesHawkularInventory.getAllScenarios().exec(FakeDataScenarios.getAllScenarios())
 
   setUp(
     setupScenario.inject(atOnceUsers(1)),
-    allScenarios.inject(rampUsers(users) over (ramp seconds))
+    loadTests.inject(rampUsers(users) over (ramp seconds))
   ).protocols(httpConf)
 }
